Updated Navbar and Guestbook Feature
Fixed navbar issue - collapse was not working properly due to navpills
class.  Apparently it is not possible to use the navbar fixed top class,
the navbar-right class, a collapse feature and nav pills all at the same
time.  It might be an option in the future to style the nav better using
CSS but time prevents it for now.  The guestbook feature has also been
updated and implemented by utilizing a handlebars template to deal with
the readout of JSON data from a Firebase database set.